feat: add rate limit protection to API calls
The Amara API will rate limit an API key based off of usage. This adds
an optional circuit breaker, that when enabled, will stop making API
calls to the Amara backend, until some time has passed. This respects
`Retry-After` headers so if the Amara API indicates a time to retry the
request we will reset the circuit breaker at that time.